requirements.txt
flask
scikit-learn
pandas
numpy
Werkzeug
psycopg2-binary